requirements.txt
RPi.GPIO
smbus2
web3
python-dotenv
